Refactor to be smarter
This is a total refactor that utilizes the puppetfile-resolver gem to do
proper dependency resolution based on the metadata.json of the local
module.

The license is also changing as part of this to be Apache-2.0 so as to
be compliant with code that is reused from puppetlabs/bolt and
glennsarti/puppetfile-resolver.

Lastly, automated gem publishing is also being added as part of this